Debug Adapter (#267)
* Renamings

* Tweaks

* Added path eq. comparer

* Small refactoring

* Existence is pain

* More elaboration

* More tweaks

* Update Source.cs

* Started adding breakpoints in API

* Added the concept of breakpoints

* Update Program.cs

* Lotta refactoring

* Fixes

* Added a separate translator

* Fix

* Working breakpoints

* Working stepping

* Fixes

* Small fix

* Added ID to stack frame and thread

* Tweaks

* Fixed thread naming

* Update Debugger.cs

* Basic locals view

* Update ProgramState.cs

* Update StackFrame.cs

* Added visualizing structured locals

* Update package.json

* Update debugadapter.ts

* Update debugadapter.ts

* Rename

* Update package.json

* Update debugadapter.ts

* Update debugadapter.ts

* Update debugadapter.ts

* Started an assets utility

* Update assets.ts

* Made use of the asset generator

* Simplified DAP capabilities

* Refactored the debug adapter

* Update DracoDebugAdapter.cs

* Added pausing

* Update Debugger.cs

* Update DracoDebugAdapter.cs

* Update DracoDebugAdapter.cs

* Added breakpoint removal

* Tweaks, correctness

* Update Method.cs

* Update DracoDebugAdapter.cs

* Almost working

* Fixed breakpoint verification

* Using cleanup

* Made thread name query platform-independent

* Update package.json

* Update ValueUtils.cs

* Update DracoDebugAdapter.cs

* let -> const

* Moved repeated code

* Removed old code

* Create tools.ts

* Update tools.ts

* Create result.ts

* Refactor

* Update tools.ts

* Added back prompt

* Added user flow function

* Update user_flow.ts

* Update user_flow.ts

* More flow

* Update user_flow.ts

* Settings cleanup

* Basic asset generation

* Added commands

* Most of the functionality is back

* Update debug_adapter.ts

* Added settings

* Create extension.ts

* Fixes

* Update language_server.ts

* Update user_flow.ts

* Update language_server.ts

* Update commands.ts

* Code cleanup

* Update publish.yaml

+++ src/Draco.Debugger/Platform/Win32PlatformMethods.cs
@@ -18,7 +18,7 @@ internal sealed class Win32PlatformMethods : IPlatformMethods
 
     private const string Kernel32 = "kernel32.dll";
 
-    private static readonly nint INVALID_HANDLE_VALUE = new nint(-1);
+    private static readonly nint INVALID_HANDLE_VALUE = new(-1);
 
     [DllImport(Kernel32, SetLastError = true)]
     private static extern nint GetStdHandle(StandardHandleType nStdHandle);
